requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
spotipy==2.18.0
urllib3==1.26.6
youtube-search==2.1.0